+++ viewer/tests/index.js
@@ -9,3 +9,4 @@ require("./inspector.test.js");
 require("./bvh.test.js");
 require("./ops.test.js");
 require("./models.test.js");
+require("./reading.test.js");